backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sentence-transformers==2.7.0
faiss-cpu==1.7.4
jieba==0.42.1
numpy==1.24.3
openai==1.6.0
httpx==0.27.2
aiofiles==23.2.1
python-docx==0.8.11
PyMuPDF==1.23.7
hf_xet
hf_transfer